Fractured code into module files
